chore(docs): next.config to mjs; rehyp-slug
