+++ packages/docs/package.json
@@ -12,13 +12,15 @@
     "@mdx-js/loader": "^1.6.22",
     "@mdx-js/react": "^1.6.22",
     "@next/mdx": "^12.0.7",
+    "clsx": "^1.1.1",
     "next": "12.0.7",
     "next-compose-plugins": "^2.2.1",
     "next-transpile-modules": "^9.0.0",
-    "nextjs-components": "^0.1.0-rc53",
+    "nextjs-components": "file:../core",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "react-live": "^2.3.0"
+    "react-live": "^2.3.0",
+    "rehype-slug": "^5.0.1"
   },
   "devDependencies": {
     "@types/react": "^17.0.38",
